+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Unreleased changes
+
+### Added
+
+- `ContractAddress.toString` function that converts the address to a string in
+  the `<index, subindex>` format.
+  
+### Fixed
+
+- Error messages in `GenericContract` now display the data, e.g., the contract
+  address, rather than `[object Object]`.
+
+
 ## 7.1.0
 
 ### Added